Log in jest test implemented
